Added all mathoperations and added getNumCommandDouble(inputs, index)

+++ src/backend/CommandNumber.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 
 public class CommandNumber extends Command {
 	
-	private Integer commandNum;
+	private Double commandNum;
 	
-	public CommandNumber(int val) {
+	public CommandNumber(double val) {
 		super("Constant", 0);
 		commandNum = val;
 	}
